pyproject.toml
[tool.poetry]
name = "Rviewer: Python & django skeleton"
version = "0.1.0"
description = ""
authors = ["rviewer-team <dev@rviewer.io>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-model-utils = "^4.3.1"
factory-boy = "^3.2.1"
psycopg2-binary = "^2.9.6"
black = "^23.3.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

